test(docker): extract setup harness
